backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0

# Core validation and settings
pydantic>=2.6.4
python-dotenv>=1.0.1
tzdata>=2024.2
email-validator>=2.2.0

# MongoDB / async data access
motor==3.3.1
pymongo==4.5.0

# Auth & security
python-jose>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.8
python-multipart>=0.0.9
requests>=2.31.0
requests-oauthlib>=2.0.0
google-auth>=2.23.0

python-dateutil>=2.8.2

